forgot to change the version lol

+++ 1.8.9-Forge/src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/Fly.java
@@ -265,7 +265,7 @@ public class Fly extends Module {
                     MovementUtils.strafe(motion);
                 } else if (verusDmged) {
                     mc.timer.timerSpeed = 1F;
-                    MovementUtils.strafe(MovementUtils.getBaseMoveSpeed());
+                    MovementUtils.strafe((float)MovementUtils.getBaseMoveSpeed());
                 }
                 break;
             case "creative":
